Separated baseline_only and baseline_only_mi
baseline_only wasn't working for Table 1 because it didn't include abuse_d_f. So, I made two versions of the data. One version is to use to create Table one (baseline_only) and the other is to impute and use for modeling (baseline_only_mi). They really only differ in that baseline_only has more variables.